feat: add library to help run experiments
BREAKING CHANGE: this changes the external interface
